[feat-planckian]: remove Tensor support for indices (reason: no backprop for indices)

+++ test/augmentation/test_augmentation.py
@@ -2695,7 +2695,7 @@ class TestPlanckianJitter:
         torch.manual_seed(0)
         input = self._get_input(device, dtype).repeat(2, 1, 1, 1)
 
-        select_from = torch.LongTensor([1, 2, 24])
+        select_from = [1, 2, 24]
         f = PlanckianJitter(select_from=select_from).to(device, dtype)
         expected = self._get_expected_output_batch(device, dtype)
         assert_close(f(input), expected, atol=1e-4, rtol=1e-5)
@@ -2704,7 +2704,7 @@ class TestPlanckianJitter:
         torch.manual_seed(0)
         input = self._get_input(device, dtype).repeat(2, 1, 1, 1)
 
-        select_from = torch.LongTensor([1, 2, 24, 3, 4, 5])
+        select_from = [1, 2, 24, 3, 4, 5]
         f = PlanckianJitter(select_from=select_from, same_on_batch=True,
                             p=1.0).to(device, dtype)
         expected = self._get_expected_output_same_on_batch(device, dtype)
